added new cvLME function
A new function "MA_cvLME_other.m" now allows to calculate the cvLME for second-level GLMs. This procedure is called from "MA_cvLME_single.m" and "MA_cvLME_multi.m" when "SPM.Sess" does not exist (which indicates that there are no sessions and that this is not a first-level GLM), so that second-level model assessment can use the usual batch functionality. As there are no sessions (as in multi-session fMRI) and no temporal order of scans (as in single-session fMRI), the present procedure uses random split-half cross-validation where MATLAB's rng is seeded with the number of scans, so that different models describing the same data use the same partition.